[refactor #10] 공통 모듈 수정
