+++ common-jpa/build.gradle
@@ -13,22 +13,27 @@ java {
 	}
 }
 
+configurations {
+	compileOnly {
+		extendsFrom annotationProcessor
+	}
+}
+
 repositories {
 	mavenCentral()
 }
 
-ext {
-	set('springCloudVersion', "2023.0.3")
+bootJar {
+	enabled = false
 }
+jar {
+	enabled = true
+}
+
 
 dependencies {
-	implementation 'org.springframework.cloud:spring-cloud-starter-netflix-eureka-server'
-}
-
-dependencyManagement {
-	imports {
-		mavenBom "org.springframework.cloud:spring-cloud-dependencies:${springCloudVersion}"
-	}
+	api 'org.springframework.boot:spring-boot-starter-data-jpa'
+	api 'org.postgresql:postgresql'
 }
 
 tasks.named('test') {
