+++ common-jpa/src/main/java/audit/BaseEntity.java
@@ -1,0 +1,44 @@
+package audit;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.MappedSuperclass;
+import java.time.LocalDateTime;
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.data.annotation.CreatedBy;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedBy;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+@Getter
+@MappedSuperclass
+@EntityListeners(AuditingEntityListener.class)
+public abstract class BaseEntity {
+    @CreatedDate
+    @Column
+    private LocalDateTime createdAt;
+
+    @CreatedBy
+    @Column
+    private String createdBy;
+
+    @LastModifiedDate
+    @Column
+    private LocalDateTime updatedAt;
+
+    @LastModifiedBy
+    @Column
+    private String updatedBy;
+
+    @Column
+    private LocalDateTime deletedAt;
+
+    @Column
+    private String deletedBy;
+
+    @Column
+    private Boolean isDeleted = false;
+
+}
