+++ common-jpa/src/main/resources/application-jpa.yml
@@ -1,0 +1,10 @@
+spring:
+  jpa:
+    properties:
+      hibernate:
+        show_sql: true
+        format_sql: true
+        use_sql_comments: true
+    generate-ddl: true
+    hibernate:
+      ddl-auto: update
