+++ common-jpa/src/main/resources/application.yml
@@ -1,0 +1,4 @@
+spring:
+  application:
+    name: common-jpa
+
