+++ common/src/main/java/cases/ErrorCase.java
@@ -1,0 +1,9 @@
+package cases;
+
+import org.springframework.http.HttpStatus;
+
+public interface ErrorCase {
+    HttpStatus getHttpStatus();
+    String getMessage();
+}
+
