+++ common/src/main/java/cases/SuccessCase.java
@@ -1,0 +1,8 @@
+package cases;
+
+import org.springframework.http.HttpStatus;
+
+public interface SuccessCase {
+    HttpStatus getHttpStatus();
+    String getMessage();
+}
