--- common/src/main/java/config/BaseEntity.java
@@ -1,47 +1,0 @@
-package config;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import java.time.LocalDateTime;
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.data.annotation.CreatedBy;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedBy;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-@Setter
-@Getter
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-public abstract class BaseEntity {
-    // Audit 필드
-    // 모든 테이블에 created_at, created_by, updated_at, updated_by, deleted_at, deleted_by 필드를 추가하여 데이터 감사 로그 기록
-    @CreatedDate
-    @Column
-    private LocalDateTime createdAt;
-
-    @CreatedBy
-    @Column
-    private String createdBy;
-
-    @LastModifiedDate
-    @Column
-    private LocalDateTime updatedAt;
-
-    @LastModifiedBy
-    @Column
-    private String updatedBy;
-
-    @Column
-    private LocalDateTime deletedAt;
-
-    @Column
-    private String deletedBy;
-
-    @Column
-    private Boolean isDeleted = false;
-
-}
