--- common/src/main/java/config/JpaAuditingConfig.java
@@ -1,13 +1,0 @@
-package config;
-
-import org.springframework.boot.autoconfigure.domain.EntityScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-@EntityScan(basePackages = "${base-package.jpa}")
-public class JpaAuditingConfig {
-
-}
-
