+++ common/src/main/java/exception/ApplicationException.java
@@ -1,5 +1,6 @@
 package exception;
 
+import cases.ErrorCase;
 import lombok.Getter;
 
 @Getter
