--- common/src/main/java/exception/ErrorCase.java
@@ -1,9 +1,0 @@
-package exception;
-
-import org.springframework.http.HttpStatus;
-
-public interface ErrorCase {
-    HttpStatus getHttpStatus();
-    String getMessage();
-}
-
