--- common/src/main/java/success/SuccessCase.java
@@ -1,8 +1,0 @@
-package success;
-
-import org.springframework.http.HttpStatus;
-
-public interface SuccessCase {
-    HttpStatus getHttpStatus();
-    String getMessage();
-}
