--- common/src/main/resources/application-common.yml
@@ -1,11 +1,0 @@
-spring:
-  jpa:
-    properties:
-      hibernate:
-        show_sql: true
-        format_sql: true
-        use_sql_comments: true
-    generate-ddl: true
-    hibernate:
-      ddl-auto: update
-
