+++ dtos/payment-dto/src/main/resources/application.yml
@@ -1,0 +1,4 @@
+spring:
+  application:
+    name: payment-dto
+
