+++ dtos/performance-dto/src/main/resources/application.yml
@@ -1,0 +1,4 @@
+spring:
+  application:
+    name: performance-dto
+
