+++ services/performance/src/main/java/com/ticketPing/performance/PerformanceApplication.java
@@ -2,10 +2,11 @@ package com.ticketPing.performance;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @SpringBootApplication
-@EnableJpaAuditing
+@ComponentScan(basePackages = {"com.ticketPing.performance", "config", "exception", "aop"})
 public class PerformanceApplication {
 
     public static void main(String[] args) {
