+++ services/performance/src/main/java/com/ticketPing/performance/presentation/response/success/PerformanceSuccessCase.java
@@ -1,9 +1,9 @@
 package com.ticketPing.performance.presentation.response.success;
 
+import cases.SuccessCase;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
-import success.SuccessCase;
 
 @Getter
 @RequiredArgsConstructor
